Refactor Committee code & unit tests
